for navmesh

+++ Assets/Scripts/PathFinding/Graph3D/Graph3DAStarMap.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using Lite;
+
 using Lite.Graph;
 
 
@@ -74,7 +74,7 @@ namespace Lite.AStar
 		}
 
 
-		public long GetGroundHeight3D(TwVector3 position)
+		public int GetGroundHeight3D(TwVector3 position)
 		{
 			int stepHeight = navGraphData.buildConfig.agentHeightStep * navGraphData.buildConfig.cellSize;
 			float posY = TwMath.mm2m(position.y);
@@ -94,9 +94,9 @@ namespace Lite.AStar
 		public Int3 TwVector3ToPoint3D(TwVector3 position)
 		{
 			int cellSize = navGraphData.buildConfig.cellSize;
-			int dx = (int)position.x - navGraphData.buildConfig.worldMinPos.x;
-			int dy = (int)position.y - navGraphData.buildConfig.worldMinPos.y;
-			int dz = (int)position.z - navGraphData.buildConfig.worldMinPos.z;
+			int dx = position.x - navGraphData.buildConfig.worldMinPos.x;
+			int dy = position.y - navGraphData.buildConfig.worldMinPos.y;
+			int dz = position.z - navGraphData.buildConfig.worldMinPos.z;
 			int x = dx / cellSize/* + ((dx % cellSize) > 0 ? 1 : 0)*/;
 			int y = dy / cellSize/* + ((dy % cellSize) > 0 ? 1 : 0)*/;
 			int z = dz / cellSize/* + ((dz % cellSize) > 0 ? 1 : 0)*/;
@@ -151,25 +151,25 @@ namespace Lite.AStar
 
 			// y = a*x + b
 			Fix64 a_xz = (Fix64)0;
-			long dx = to.x - from.x;
-			long dz = to.z - from.z;
+			int dx = to.x - from.x;
+			int dz = to.z - from.z;
 			if (Math.Abs(dx) > Math.Abs(dz))
 			{
 				a_xz = (Fix64)dz / (Fix64)dx;
 				int step = to.x - from.x > 0 ? stepLen : -stepLen;
-				long lastY = from.y;
-				for (long x = from.x + step; step > 0 ? x < to.x + step : x > to.x - step; x += step)
+				int lastY = from.y;
+				for (int x = from.x + step; step > 0 ? x < to.x + step : x > to.x - step; x += step)
 				{
 					x = step > 0 ? System.Math.Min(x, to.x) : System.Math.Max(x, to.x);
 					Fix64 z = (Fix64)from.z + a_xz * (Fix64)(x - from.x);
-					long y = lastY;
+					int y = lastY;
 
 					// stairs
 					bool passable = false;
 					for (int iy = halfAgentHeightStep; iy >= -halfAgentHeightStep; iy--)
 					{
-						long tmpy = y + iy * navGraphData.buildConfig.cellSize;
-						Int3 pt3d = TwVector3ToPoint3D(new TwVector3(x, tmpy, (long)z));
+						int tmpy = y + iy * navGraphData.buildConfig.cellSize;
+						Int3 pt3d = TwVector3ToPoint3D(new TwVector3(x, tmpy, (int)z));
 						var node = GetNodeAt(pt3d.x, pt3d.y, pt3d.z);
 						if (IsNodePassable(pt3d.x, pt3d.y, pt3d.z))
 						{
@@ -191,26 +191,26 @@ namespace Lite.AStar
 						break;
 					}*/
 
-					blockPoint.Set(x, (long)y, (long)z);
+					blockPoint.Set(x, (int)y, (int)z);
 				}
 			}
 			else
 			{
 				a_xz = (Fix64)dx / (Fix64)dz;
 				int step = to.z - from.z > 0 ? stepLen : -stepLen;
-				long lastY = from.y;
-				for (long z = from.z + step; step > 0 ? z < to.z + step : z > to.z - step; z += step)
+				int lastY = from.y;
+				for (int z = from.z + step; step > 0 ? z < to.z + step : z > to.z - step; z += step)
 				{
 					z = step > 0 ? System.Math.Min(z, to.z) : System.Math.Max(z, to.z);
 					Fix64 x = (Fix64)from.x + a_xz * (Fix64)(z - from.z);
-					long y = lastY;
+					int y = lastY;
 
 					// stairs
 					bool passable = false;
 					for (int iy = halfAgentHeightStep; iy >= -halfAgentHeightStep; iy--)
 					{
-						long tmpy = y + iy * navGraphData.buildConfig.cellSize;
-						Int3 pt3d = TwVector3ToPoint3D(new TwVector3((long)x, tmpy, z));
+						int tmpy = y + iy * navGraphData.buildConfig.cellSize;
+						Int3 pt3d = TwVector3ToPoint3D(new TwVector3((int)x, tmpy, z));
 						var node = GetNodeAt(pt3d.x, pt3d.y, pt3d.z);
 						if (IsNodePassable(pt3d.x, pt3d.y, pt3d.z))
 						{
@@ -232,7 +232,7 @@ namespace Lite.AStar
 						break;
 					}*/
 
-					blockPoint.Set((long)x, (long)y, z);
+					blockPoint.Set((int)x, (int)y, z);
 				}
 			}
